現在地が取得できないけど、一旦コミット。
