+++ app/src/main/res/layout/fragment_map.xml
@@ -1,7 +1,12 @@
-<RelativeLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:paddingLeft="@dimen/activity_horizontal_margin"
+    android:paddingRight="@dimen/activity_horizontal_margin"
+    android:paddingTop="@dimen/activity_vertical_margin"
+    android:paddingBottom="@dimen/activity_vertical_margin"
+    tools:context="jp.co.icomsys.it21.icomradar.MainActivity$RadarMapFragment">
 
     <com.google.android.gms.maps.MapView
         android:id="@+id/map"
@@ -9,11 +14,12 @@
         android:layout_height="match_parent" />
 
     <ImageView
-        android:id="@+id/icomradar"
+        android:id="@+id/imageView"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:scaleType="centerCrop"
-        android:contentDescription="icom radar"
-        android:src="@drawable/radar" />
+        android:scaleType="fitXY"
+        android:src="@drawable/icomrador"
+        android:layout_alignParentRight="true"
+        android:layout_alignParentEnd="true" />
 
 </RelativeLayout>
